+++ composeApp/src/commonMain/kotlin/tech/kotlinlang/permission/ui/component/UiComponentsScreen.kt
@@ -4,6 +4,8 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
@@ -71,10 +73,12 @@ private fun PdfViewerContent(modifier: Modifier) {
         contentAlignment = Alignment.Center,
     ) {
         val pdfViewerState by rememberPdfViewerState("https://kotlinlang.org/docs/kotlin-reference.pdf")
-        val currentPdfViewerState = pdfViewerState
-        if (currentPdfViewerState is PdfViewerState.Success) {
-            PdfViewer(currentPdfViewerState)
-        }
+        PdfViewer(
+            modifier = Modifier
+                .fillMaxWidth()
+                .aspectRatio(1F),
+            pdfViewerState = pdfViewerState
+        )
     }
 }