+++ pdfviewer/src/commonMain/kotlin/tech/kotlinlang/pdfviewer/PdfViewer.kt
@@ -1,18 +1,31 @@
 package tech.kotlinlang.pdfviewer
 
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun PdfViewer(pdfViewerState: PdfViewerState.Success) {
-    PdfViewerLoaded(
-        modifier = Modifier
-            .fillMaxWidth()
-            .aspectRatio(1F),
-        bytes = pdfViewerState.bytes,
-    )
+fun PdfViewer(
+    modifier: Modifier = Modifier,
+    pdfViewerState: PdfViewerState
+) {
+    if (pdfViewerState is PdfViewerState.Success) {
+        PdfViewerLoaded(
+            modifier = modifier,
+            bytes = pdfViewerState.bytes,
+        )
+    } else {
+        Box(
+            modifier = modifier,
+            contentAlignment = Alignment.Center,
+        ) {
+            if (pdfViewerState is PdfViewerState.Loading) {
+                CircularProgressIndicator()
+            }
+        }
+    }
 }
 
 @Composable